requirements.txt
absl-py==0.10.0
jax==0.1.76
jaxlib==0.1.55
matplotlib==3.3.1
numpy==1.19.2
scikit-learn==0.23.2
scipy==1.5.2
statsmodels==0.12.0
tqdm==4.49.0
